utilities/requirements.txt
boto3==1.26.76
botocore==1.29.165
click==8.1.3
fiftyone==0.24.1
fiftyone_brain==0.16.1
fiftyone_db==1.1.4
matplotlib==3.7.4
numpy==1.24.3
opencv_python==4.9.0.80
opencv_python_headless==4.5.5.64
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.4.0
pycryptodome==3.10.1
pytest==8.3.1
scikit_learn==1.3.2
setuptools==68.2.2
torch==2.3.1
tqdm==4.64.1
urllib3==1.26.18
.\requirements\deeplake-3.8.8+asus.tar.gz
.\requirements\aasdatahub-0.6.2-cp38-abi3-win_amd64.whl
.\requirements\aisdatahub-0.5.0+20231220-cp38-abi3-win_amd64.whl
